requirements.txt
# Core data science and ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning frameworks
tensorflow>=2.8.0
torch>=1.11.0
transformers>=4.15.0

# Financial data and analysis
yfinance>=0.1.87
fredapi>=0.4.3
quandl>=3.7.0
alpha-vantage>=2.3.1

# Time series analysis
statsmodels>=0.13.0
arch>=5.3.0
pmdarima>=1.8.5

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.6.0
bokeh>=2.4.0

# Explainability and interpretability
shap>=0.41.0
lime>=0.2.0.1
eli5>=0.11.0

# Data processing and utilities
tqdm>=4.62.0
joblib>=1.1.0
openpyxl>=3.0.9
xlrd>=2.0.1

# Configuration and logging
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0
hydra-core>=1.1.0

# Jupyter and notebook extensions
jupyter>=1.0.0
jupyterlab>=3.2.0
ipywidgets>=7.6.0
notebook>=6.4.0

# Code quality and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.931

# Development and documentation
pre-commit>=2.17.0
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0

# Model deployment and serving
fastapi>=0.75.0
uvicorn>=0.17.0
streamlit>=1.8.0

# Experiment tracking and MLOps
mlflow>=1.24.0
wandb>=0.12.0
tensorboard>=2.8.0

# Additional utilities
requests>=2.27.0
beautifulsoup4>=4.10.0
lxml>=4.7.0 